+++ CHANGELOG.md
@@ -24,6 +24,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Added private function Update-RubrikuserRole to handle the addition and removal of permissions for the various roles.
 * Addresses [Issue 108](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/108)
 
+## 2019-09-26
+
+### Added initial support for VMware vCloud Director (vCD) API endpoints in CDM
+
+* Added cmdlets `Update-RubrikVCD`, `Set-RubrikVCD`, `Restore-RubrikVApp`, `Protect-RubrikVApp`, `Get-RubrikVcdTemplateExportOptions`, `Get-RubrikVCD`, `Get-RubrikVappSnapshot`, `RubrikVAppRecoverOptions`, `Get-RubrikVAppExportOptions`, `Get-RubrikVApp`, `Export-RubrikVCDTemplate`, `Export-RubrikVApp` and related tests as requested in [Issue 273](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/273)
+* Updated `Get-RubrikSnapshot` to support vCD vApps
+* Misc bug fixes and typos corrected
+
 ## 2019-09-24
 
 ### Added [Private Function to apply TypeName information to objects]
@@ -58,6 +66,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Addresses [Issue 348](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/348)
 
 ## 2019-09-09
+
+### Added [Get-RubrikObject]
+
+* Created Get-RubrikObject cmdlet and respective Unit test.
+* Addresses [Issue 349](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/349)
 
 ### Added [Various Unit Tests]